+++ examples/2_Audio/control/PluckedStringMidi/PluckedStringMidi.pde
@@ -37,7 +37,7 @@ ModelRenderer renderer;
 float speed = 0;
 float pos = 100;
 
-ArrayList<MidiControler> midiCtrls = new ArrayList<MidiControler>();
+ArrayList<MidiController> midiCtrls = new ArrayList<MidiController>();
 ArrayList<midiNote> midiNotes = new ArrayList<midiNote>();
 ///////////////////////////////////////
 
@@ -69,10 +69,10 @@ void setup()
   myBus = new MidiBus(this, 0, 1); // Create a new MidiBus with no input device and the default Java Sound Synthesizer as the output device.
 
 
-  midiCtrls.add(MidiControler.addMidiControler(simUGen.mdl,1, 0.01, 0.3, "sprd", "stiffness",0.05));
-  midiCtrls.add(MidiControler.addMidiControler(simUGen.mdl,2, 0.0001, 0.1, "sprd", "damping",0.05));
-  midiCtrls.add(MidiControler.addMidiControler(simUGen.mdl,3, 0.5, 1.5, "str", "mass",0.05));
-  midiCtrls.add(MidiControler.addMidiControler(simUGen.mdl,4, 20, 100, "gnd", "distX",0.05));
+  midiCtrls.add(MidiController.addMidiController(simUGen.mdl,1, 0.01, 0.3, "sprd", "stiffness",0.05));
+  midiCtrls.add(MidiController.addMidiController(simUGen.mdl,2, 0.0001, 0.1, "sprd", "damping",0.05));
+  midiCtrls.add(MidiController.addMidiController(simUGen.mdl,3, 0.5, 1.5, "str", "mass",0.05));
+  midiCtrls.add(MidiController.addMidiController(simUGen.mdl,4, 20, 100, "gnd", "distX",0.05));
 
   //midiNotes.add(new midiNote(0.01, 0.9, "str",0,simUGen.nbmass - 1,"Y","impulse"));
   midiNotes.add(new midiNote(0.1, 10, "str",0,simUGen.nbmass - 1,"Y","pluck"));
@@ -164,7 +164,7 @@ void keyReleased() {
 void controllerChange(int channel, int number, int value) {
   synchronized(lock)
   {
-    for (MidiControler mc : midiCtrls)
+    for (MidiController mc : midiCtrls)
     {
       mc.changeParam(number, value);
     }
